fix: ...I think I fixed the CI.
The issue - Github Workflow would fail during the UPM branching stage. This turned out to be due to `git commit -am` which returns nonzero when you attempt to make an empty commit (a commit with no changes). Apparently if a function returns nonzero, the CI immediately fails.

The fix - Add `--allow-empty` to all commits which might be empty